Updated AuthorProfile page to use common HeroHeader
