+++ .gitignore
@@ -43,3 +43,4 @@ next-env.d.ts
 external_references/
 
 .cursor/
+.claude/
