+++ app/author/[id]/layout.tsx
@@ -1,9 +1,3 @@
-import { PageLayout } from '@/app/layouts/PageLayout';
-
 export default function AuthorProfileLayout({ children }: { children: React.ReactNode }) {
-  return (
-    <PageLayout rightSidebar={null}>
-      <div className="w-full">{children}</div>
-    </PageLayout>
-  );
+  return <>{children}</>;
 }
